perf(ui): use Async component to avoid loading ressources - WF-107
