+++ src/ui/src/components/core/content/CorePlotlyGraph.vue
@@ -10,49 +10,11 @@ For example, implement cross-filtering.
 </template>
 
 <script lang="ts">
-import { FieldType } from "@/writerTypes";
-import { cssClasses } from "@/renderer/sharedStyleFields";
 import { WdsColor } from "@/wds/tokens";
 
-const description = "A component that displays Plotly graphs.";
+import definition from "./CorePlotlyGraph.def";
 
-const defaultSpec = {
-	data: [
-		{
-			x: ["a", "b", "c"],
-			y: [22, 25, 29],
-			type: "bar",
-		},
-	],
-};
-
-export default {
-	writer: {
-		name: "Plotly Graph",
-		description,
-		category: "Content",
-		fields: {
-			spec: {
-				name: "Graph specification",
-				default: JSON.stringify(defaultSpec, null, 2),
-				desc: "Plotly graph specification. Pass it using state, e.g. @{fig}, or paste a JSON specification.",
-				type: FieldType.Object,
-			},
-			cssClasses,
-		},
-		events: {
-			"plotly-click": {
-				desc: "Sends a list with the clicked points.",
-			},
-			"plotly-selected": {
-				desc: "Sends a list with the selected points.",
-			},
-			"plotly-deselect": {
-				desc: "Triggered when points are deselected.",
-			},
-		},
-	},
-};
+export default { writer: definition };
 </script>
 
 <script setup lang="ts">
