+++ src/ui/src/components/core/embed/CorePDF.vue
@@ -45,67 +45,15 @@
 </template>
 
 <script lang="ts">
-import { FieldType } from "@/writerTypes";
-import {
-	cssClasses,
-	separatorColor,
-	primaryTextColor,
-	containerBackgroundColor,
-} from "@/renderer/sharedStyleFields";
-import WdsControl from "@/wds/WdsControl.vue";
+import definition from "./CorePDF.def";
 
-const description = "A component to embed PDF documents.";
-
-export default {
-	writer: {
-		name: "PDF",
-		description,
-		category: "Embed",
-		fields: {
-			source: {
-				name: "PDF source",
-				type: FieldType.Text,
-				desc: "A valid URL. Alternatively, you can provide a state reference to a packed PDF file.",
-			},
-			highlights: {
-				name: "Highlights",
-				default: JSON.stringify([]),
-				desc: "A list of highlights to be applied to the PDF as a JSON array of strings.",
-				type: FieldType.Object,
-			},
-			selectedMatch: {
-				name: "Selected highlight match",
-				default: null,
-				desc: "The index of the selected highlight match.",
-				type: FieldType.Number,
-			},
-			page: {
-				name: "Page",
-				type: FieldType.Number,
-				desc: "The page to be displayed.",
-			},
-			controls: {
-				name: "Controls",
-				type: FieldType.Text,
-				options: {
-					yes: "Yes",
-					no: "No",
-				},
-				desc: "Show controls to navigate the PDF.",
-				default: "yes",
-			},
-			containerBackgroundColor,
-			separatorColor,
-			primaryTextColor,
-			cssClasses,
-		},
-	},
-};
+export default { writer: definition };
 </script>
 
 <script setup lang="ts">
 import { inject, ref, watch, computed, onMounted } from "vue";
 import injectionKeys from "@/injectionKeys";
+import WdsControl from "@/wds/WdsControl.vue";
 import "@tato30/vue-pdf/style.css";
 
 type MatchType = { str: string; page: number; index: number };
